Code polish
