+++ public/locales/he/translation.json
@@ -100,7 +100,7 @@
     "connectWallet": {
       "modalTitle": "חבר ארנק",
       "actionButton": "קראתי ומסכים",
-      "content1": "על ידי חיבור הארנק שלי, אני מסכים ל<1>תנאי השימוש</1> ול<2>מדיניות הקוקיז והפרטיות</2> של האתר הזה.",
+      "content1": "על ידי חיבור הארנק שלי, אני מסכים ל<0>תנאי השימוש</0> ול<1>מדיניות הקוקיז והפרטיות</1> של האתר הזה.",
       "error1": "{{error}} שגיאה:"
     },
     "confirm": {
