+++ src/components/strategies/overview/utils.ts
@@ -1,6 +1,7 @@
 import BigNumber from 'bignumber.js';
 import { Strategy } from 'libs/queries';
-import { StrategySort } from './StrategyFilterSort';
+import { TFunction } from 'libs/translations';
+import { StrategyFilter, StrategySort } from './StrategyFilterSort';
 
 export const getCompareFunctionBySortType = (sortType: StrategySort) => {
   let firstPairComparison: number;
@@ -32,4 +33,44 @@ export const getCompareFunctionBySortType = (sortType: StrategySort) => {
       return (a: Strategy, b: Strategy) =>
         new BigNumber(a.idDisplay).minus(b.idDisplay).times(-1).toNumber();
   }
+};
+
+export const getSortAndFilterItems = (
+  t: TFunction<string, undefined, string>
+) => {
+  const sortItems = [
+    {
+      title: t('pages.strategyOverview.header.filterMenu.items.item1'),
+      item: StrategySort.Recent,
+    },
+    {
+      title: t('pages.strategyOverview.header.filterMenu.items.item2'),
+      item: StrategySort.Old,
+    },
+    {
+      title: t('pages.strategyOverview.header.filterMenu.items.item3'),
+      item: StrategySort.PairAscending,
+    },
+    {
+      title: t('pages.strategyOverview.header.filterMenu.items.item4'),
+      item: StrategySort.PairDescending,
+    },
+  ];
+
+  const filterItems = [
+    {
+      title: t('pages.strategyOverview.header.filterMenu.items.item5'),
+      item: StrategyFilter.All,
+    },
+    {
+      title: t('pages.strategyOverview.header.filterMenu.items.item6'),
+      item: StrategyFilter.Active,
+    },
+    {
+      title: t('pages.strategyOverview.header.filterMenu.items.item7'),
+      item: StrategyFilter.Inactive,
+    },
+  ];
+
+  return { sortItems, filterItems };
 };
